# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 公共空间反馈处理追踪 API
  version: 1.0.0
  description: 面向提交者、处理人员、管理员与审计人员的离线可运行反馈服务。
servers:
  - url: http://localhost:8080/api/v1
paths:
  /feedback:
    post:
      summary: 提交公共空间反馈
      operationId: submitFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitFeedback' }
      responses:
        '201': { description: 已生成受理编号与查询令牌 }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  schemas:
    SubmitFeedback:
      type: object
      required: [area_id, facility_category_id, subject_code, title, description, location]
      properties:
        area_id: { type: string }
        facility_category_id: { type: string }
        subject_code: { type: string }
        priority: { type: string, enum: [low, normal, high, critical] }
        title: { type: string, minLength: 5, maxLength: 120 }
        description: { type: string, minLength: 10, maxLength: 4000 }
        location: { type: string, maxLength: 300 }
        anonymous: { type: boolean }
        submitter_name: { type: string }
        submitter_contact: { type: string }
...
    Error:
      type: object
      required: [code, message, field_errors, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        field_errors: { type: array, items: { type: object } }
        request_id: { type: string }
  responses:
    ValidationError:
      description: 请求字段或业务状态不合法
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...